requirements.txt
Django==1.8.3

django-filter==0.11.0

djangorestframework==3.2.3

djangorestframework-filters==0.4.0

dj-database-url==0.3.0

dj-static==0.0.6

gunicorn==19.1.1

static==0.4

wsgiref==0.1.2

psycopg2==2.5.1

requests==2.7.0